added some features

+++ src/main/java/com/luinwee/restaurant/dto/ProductRequest.java
@@ -2,7 +2,7 @@ package com.luinwee.restaurant.dto;
 
 public record ProductRequest(
         Long productId,
-        float amount
+        int amount
 
 ) {
 }
